requirements.txt
streamlit
praw
python-dotenv